[MS] Moved devices functions to their own file
